MBS-10577 Extract plugins check from plugin
